Refactor editor into component and size writing area
